Streamline msgParser.js into focused modules
Decomposed the 453-line msgParser.js into 7 focused modules (each under 200 lines) to improve code maintainability and readability:

- msgParser.js (48 lines) - Main coordinator
- formatters.js (65 lines) - Address and date formatting
- bodyExtractor.js (73 lines) - Body and attachment extraction
- textUtils.js (79 lines) - Text cleaning and flexible date parsing
- boundaryDetector.js (51 lines) - Message boundary detection
- emailExtractor.js (117 lines) - Metadata extraction from sections
- chainParser.js (52 lines) - Chain detection coordinator

Original file backed up to deprecated/msgParser.js.backup.
All functionality preserved - no behavioral changes.
