+++ index.html
@@ -13,7 +13,7 @@
         <header class="app-header">
             <h1>QuickChain</h1>
             <div class="header-controls">
-                <div class="commit-date">Commit date: 03:04AM on November 13, 2025</div>
+                <div class="commit-date">Commit date: 03:40AM on November 13, 2025</div>
                 <button id="clearAllBtn" class="clear-btn">Clear All</button>
             </div>
         </header>
